Introduced AbstractRenderLayer into RenderLayer heirarchy.
