+++ Artemis.Engine/Graphics/RenderLayer.cs
@@ -16,55 +16,20 @@ using System.Linq;
 
 namespace Artemis.Engine.Graphics
 {
-    public class RenderLayer : UriTreeObserverNode<RenderLayer, RenderableGroup>
+    public class RenderLayer : AbstractRenderLayer<RenderLayer>
     {
-        private const string TOP_LEVEL = "ALL"; // The name of the top level of renderable objects.
         private RenderPipeline rp; // the global render pipeline,
         private World _world;
         private AbstractCamera _camera;
         private bool _requiresTargetTransformRecalc; // whether or not we need to recalculate the TargetTransform 
                                                      // matrix (when target resolution changes for example).
-        private bool midRender; // whether or not we've entered a render cycle.
+        private Matrix _targetTransform;
+        private Predicate<RenderableObject> isVisibleToCameraPredicate;
 
         private GlobalLayerScaleType _layerScaleType;
         private UniformLayerScaleType _uniformScaleType;
         private GlobalLayerScaleType? _newLayerScaleType;
         private UniformLayerScaleType? _newUniformScaleType;
-
-        private readonly List<RenderableToAdd> renderablesToAdd = new List<RenderableToAdd>();
-
-        // the list of the renderables we've seen so far in this render cycle.
-        private readonly HashSet<RenderableObject> seenRenderables = new HashSet<RenderableObject>();
-
-        // the list of renderable groups we've seen so far in this render cycle.
-        private readonly HashSet<RenderableGroup> seenGroups = new HashSet<RenderableGroup>(); 
-
-        internal Matrix _targetTransform;
-        internal string tempFullName { get; private set; }
-
-        private struct RenderableToAdd
-        {
-            public RenderableObject Object;
-            public string Name;
-            public bool Anonymous;
-        }
-
-        /// <summary>
-        /// A delegate defining what to do upon encountering a renderable
-        /// object in a renderable group.
-        /// </summary>
-        /// <param name="?"></param>
-        public delegate void RenderAction(RenderableObject obj);
-
-        /// <summary>
-        /// The target we're rendering to.
-        /// </summary>
-        protected RenderTarget2D LayerTarget { get; private set; }
-
-        /// <summary>
-        /// The top level renderable group.
-        /// </summary>
-        protected RenderableGroup AllRenderables { get; private set; }
 
         /// <summary>
         /// Determines whether or not the layer is scaled "Uniformly" (the entire layer is
@@ -80,7 +45,7 @@ namespace Artemis.Engine.Graphics
                     return;
                 _layerScaleType = value;
 
-                if (midRender)
+                if (MidRender)
                 {
                     _newLayerScaleType = value;
                     _requiresTargetTransformRecalc = true;
@@ -103,7 +68,7 @@ namespace Artemis.Engine.Graphics
             {
                 if (value == _uniformScaleType)
                     return;               
-                if (midRender)
+                if (MidRender)
                 {
                     _newUniformScaleType = value;
                     _requiresTargetTransformRecalc = true;
@@ -117,14 +82,6 @@ namespace Artemis.Engine.Graphics
         }
 
         /// <summary>
-        /// The global RenderGroupOptions, determining whether or not the top level items should be 
-        /// rendered first or the subnodes first. The default value is "RenderOrder.RenderGroupOptions.AllPre".
-        /// 
-        /// Note: If RenderOrder is not null, this value is not used.
-        /// </summary>
-        public RenderOrder.RenderGroupOptions GlobalRenderGroupOptions;
-
-        /// <summary>
         /// The camera attached to this layer.
         /// </summary>
         public AbstractCamera Camera
@@ -132,7 +89,7 @@ namespace Artemis.Engine.Graphics
             get { return _camera; }
             set
             {
-                if (midRender)
+                if (MidRender)
                     throw new CameraException(
                         String.Format(
                             "Cannot set Camera on render layer with name '{0}' " +
@@ -154,34 +111,6 @@ namespace Artemis.Engine.Graphics
             get { return _world; }
         }
 
-        /// <summary>
-        /// Whether or not we are in the middle of rendering this layer.
-        /// </summary>
-        public bool MidRender { get { return midRender; } }
-
-        /// <summary>
-        /// Whether or not this render layer is managed by a LayerManager or not.
-        /// </summary>
-        public bool Managed { get; internal set; }
-
-        /// <summary>
-        /// The order in which objects get rendered.
-        /// </summary>
-        public RenderOrder RenderOrder { get; private set; }
-
-        /// <summary>
-        /// THe list of actions representing the render order.
-        /// </summary>
-        public List<RenderOrder.AbstractRenderOrderAction> RenderOrderActions
-        {
-            get
-            {
-                if (RenderOrder == null)
-                    return new List<RenderOrder.AbstractRenderOrderAction>();
-                return RenderOrder.Actions;
-            }
-        }
-
         public RenderLayer(string fullName)
             : this(fullName, new NullCamera()) { }
 
@@ -200,17 +129,6 @@ namespace Artemis.Engine.Graphics
         {
             rp = ArtemisEngine.RenderPipeline; // for convenience
 
-            // We have to store the full name until the layer gets added to 
-            // a LayerManager.
-            tempFullName = fullName;
-            Managed = false;
-
-            AllRenderables = new RenderableGroup(TOP_LEVEL);
-            AddObservedNode(TOP_LEVEL, AllRenderables);
-
-            LayerTarget = ArtemisEngine.RenderPipeline.CreateRenderTarget();
-
-            GlobalRenderGroupOptions = RenderOrder.RenderGroupOptions.AllPre;
             LayerScaleType = layerScaleType;
             UniformScaleType = uniformScaleType;
             RecalculateTargetTransform();
@@ -219,176 +137,6 @@ namespace Artemis.Engine.Graphics
             camera.Layer = this;
 
             _world = world;
-        }
-
-        /// <summary>
-        /// Add an empty group with the given name to the layer.
-        /// </summary>
-        /// <param name="name"></param>
-        public void AddGroup(string name)
-        {
-            AllRenderables.AddSubnode(name, new RenderableGroup(UriUtilities.GetLastPart(name)));
-        }
-
-        /// <summary>
-        /// Add the renderable group to the layer.
-        /// </summary>
-        /// <param name="group"></param>
-        public void AddGroup(RenderableGroup group)
-        {
-            AllRenderables.AddSubnode(group.Name, group);
-        }
-
-        /// <summary>
-        /// Add the renderable group with the given name to the layer.
-        /// </summary>
-        /// <param name="name"></param>
-        /// <param name="group"></param>
-        public void AddGroup(string name, RenderableGroup group)
-        {
-            AllRenderables.AddSubnode(name, group);
-        }
-
-        /// <summary>
-        /// Add a RenderableObject to this render layer.
-        /// </summary>
-        /// <param name="name"></param>
-        /// <param name="item"></param>
-        public void AddItem(string name, RenderableObject item)
-        {
-            if (midRender)
-                renderablesToAdd.Add(new RenderableToAdd { Name = name, Object = item });
-            else
-                AllRenderables.InsertItem(name, item);
-        }
-
-        /// <summary>
-        /// Add a Form to this render layer.
-        /// </summary>
-        /// <param name="form"></param>
-        public void AddItem(Form form)
-        {
-            AddItem(form.Name, form);
-        }
-
-        /// <summary>
-        /// Add an anonymous item to this layer.
-        /// </summary>
-        /// <param name="item"></param>
-        public void AddAnonymousItem(RenderableObject item)
-        {
-            if (midRender)
-                renderablesToAdd.Add(new RenderableToAdd { Name = null, Object = item, Anonymous = true });
-            else
-                AllRenderables.AddAnonymousItem(item);
-        }
-
-        /// <summary>
-        /// Add an anonymous item to the renderable group with the given name in this layer.
-        /// </summary>
-        /// <param name="groupName"></param>
-        /// <param name="item"></param>
-        public void AddAnonymousItem(string groupName, RenderableObject item)
-        {
-            if (midRender)
-                renderablesToAdd.Add(new RenderableToAdd { Name = groupName, Object = item, Anonymous = true });
-            else
-                AllRenderables.AddAnonymousItem(groupName, item);
-        }
-
-        /// <summary>
-        /// Add an anonymous form to this layer.
-        /// </summary>
-        /// <param name="form"></param>
-        public void AddAnonymousItem(Form form)
-        {
-            AddAnonymousItem((RenderableObject)form);
-        }
-
-        /// <summary>
-        /// Add an anonymous form to the renderable group with the given name in this layer.
-        /// </summary>
-        /// <param name="groupName"></param>
-        /// <param name="form"></param>
-        public void AddAnonymousItem(string groupName, Form form)
-        {
-            AddAnonymousItem(groupName, (RenderableObject)form);
-        }
-
-        /// <summary>
-        /// Set the layer's RenderOrder to the render the groups/items with the given names.
-        /// </summary>
-        /// <param name="type">The type that each name represents (item or group).</param>
-        /// <param name="names"></param>
-        public void SetRenderOrder(RenderOrder.RenderType type, params string[] names)
-        {
-            var renderOrder = new RenderOrder();
-            switch (type)
-            {
-                case RenderOrder.RenderType.Item:
-                    foreach (var name in names)
-                        renderOrder.AddRenderItem(name);
-                    break;
-                case RenderOrder.RenderType.Group:
-                    foreach (var name in names)
-                        renderOrder.AddRenderGroup(name);
-                    break;
-                default:
-                    throw new RenderOrderException(
-                        String.Format(
-                            "Can't handle RenderOrder.RenderType '{0}'" +
-                            " for more options see the other overloads of `SetRenderOrder`.", type
-                            )
-                        );
-            }
-            SetRenderOrder(renderOrder);
-        }
-
-        /// <summary>
-        /// Set the render order to the render the given list of groups/items with the given
-        /// names and types.
-        /// </summary>
-        /// <param name="types"></param>
-        /// <param name="names"></param>
-        public void SetRenderOrder(RenderOrder.RenderType[] types, string[] names)
-        {
-            if (types.Length != names.Length)
-                throw new RenderOrderException(
-                    "The length of the given types array must match the length of the given names array.");
-
-            var renderOrder = new RenderOrder();
-            for (int i = 0; i < types.Length; i++)
-            {
-                if (types[i] == RenderOrder.RenderType.Item)
-                    renderOrder.AddRenderItem(names[i]);
-                else if (types[i] == RenderOrder.RenderType.Group)
-                    renderOrder.AddRenderGroup(names[i]);
-            }
-            SetRenderOrder(renderOrder);
-        }
-
-        /// <summary>
-        /// Set the render order to the given list of AbstractRenderOrderActions.
-        /// </summary>
-        /// <param name="actions"></param>
-        public void SetRenderOrder(params RenderOrder.AbstractRenderOrderAction[] actions)
-        {
-            SetRenderOrder(new RenderOrder(actions.ToList()));
-        }
-
-        /// <summary>
-        /// Set the render order to the given RenderOrder object.
-        /// </summary>
-        /// <param name="order"></param>
-        public void SetRenderOrder(RenderOrder order)
-        {
-            if (midRender)
-                throw new RenderOrderException(
-                    String.Format(
-                        "Cannot set render order on layer '{0}' during the render cycle.", FullName));
-            RenderOrder = order;
-
-            // Calculate render order.
         }
 
         /// <summary>
@@ -528,7 +276,7 @@ namespace Artemis.Engine.Graphics
             // Swap out the original scale for the newly calculated scale.
             obj.RenderComponents.Scale = resultingScale;
 
-            obj.InternalRender(seenRenderables);
+            obj.InternalRender(SeenRenderables);
 
             obj.RenderComponents.Scale = hasOriginalScale ? (Vector2?)originalScale : null;
         }
@@ -540,30 +288,30 @@ namespace Artemis.Engine.Graphics
         /// For RenderLayer, the render action returned depends on the LayerScaleType.
         /// </summary>
         /// <returns></returns>
-        private RenderAction GetRenderAction(Predicate<RenderableObject> isVisibileToCamera)
+        protected override RenderableHandler GetRenderAction()
         {
             switch (LayerScaleType)
             {
                 case GlobalLayerScaleType.Uniform:
-                    if (isVisibileToCamera == null)
-                        return obj => obj.InternalRender(seenRenderables);
+                    if (isVisibleToCameraPredicate == null)
+                        return obj => obj.InternalRender(SeenRenderables);
                     else
                         return obj =>
                         {
-                            if (isVisibileToCamera(obj))
-                                obj.InternalRender(seenRenderables);
+                            if (isVisibleToCameraPredicate(obj))
+                                obj.InternalRender(SeenRenderables);
                         };
                 case GlobalLayerScaleType.Dynamic:
                     var isBaseRes = ArtemisEngine.DisplayManager.IsBaseResolution;
                     var crntRes   = ArtemisEngine.DisplayManager.WindowResolution;
                     var resScale  = ArtemisEngine.DisplayManager.ResolutionScale;
 
-                    if (isVisibileToCamera == null)
+                    if (isVisibleToCameraPredicate == null)
                         return obj => ProcessDynamicallyScaledRenderable(obj, isBaseRes, crntRes, resScale);
                     else
                         return obj =>
                         {
-                            if (isVisibileToCamera(obj))
+                            if (isVisibleToCameraPredicate(obj))
                                 ProcessDynamicallyScaledRenderable(obj, isBaseRes, crntRes, resScale);
                         };
                 default:
@@ -596,26 +344,8 @@ namespace Artemis.Engine.Graphics
             return renderables;
         }
 
-        /// <summary>
-        /// Render all items and sublayers on this layer.
-        /// </summary>
-        public void RenderAll()
+        protected override void PreRender()
         {
-            foreach (var layer in Subnodes.Values)
-            {
-                layer.RenderAll();
-            }
-
-            Render();
-        }
-
-        /// <summary>
-        /// Render all items in this layer.
-        /// </summary>
-        public void Render()
-        {
-            midRender = true;
-
             // Reset the RenderTarget if the resolution has changed.
             if (ArtemisEngine.DisplayManager.ResolutionChanged)
             {
@@ -629,77 +359,30 @@ namespace Artemis.Engine.Graphics
                 RecalculateTargetTransform();
             }
 
-            // Set the RenderPipeline's target to be this layer's target, clear it,
-            // and lock the matrix property so any user matrix transformations are applied
-            // to the Camera's matrix.
-            rp.ClearRenderProperties();
-
-            rp.SetRenderTarget(LayerTarget);
-            rp.ClearGraphicsDevice(Color.Transparent);
-
-            rp.SetRenderProperties(m: Camera.WorldToTargetTransform);
-            rp.LockMatrix();
-
             // Get all the renderable objects.
             var renderables = GetCameraVisibleRenderables();
 
             // Create the predicate that checks if a given renderable is visible.
-            Predicate<RenderableObject> isVisibleToCamera = null; // if null, then every renderable is visible.
+            isVisibleToCameraPredicate = null; // if null, then every renderable is visible.
             if (!(Camera is NullCamera))
             {
                 var hashSet = (HashSet<RenderableObject>)renderables;
-                isVisibleToCamera = obj => hashSet.Contains(obj);
+                isVisibleToCameraPredicate = obj => hashSet.Contains(obj);
             }
+        }
+    
+        protected override void PostSetupLayerTarget()
+        {
+            rp.SetRenderProperties(m: Camera.WorldToTargetTransform);
+            rp.LockMatrix();
+        }
 
-            // Get the render action for each encountered RenderableObject.
-            RenderAction renderAction = GetRenderAction(isVisibleToCamera);
-
-            seenRenderables.Clear();
-            seenGroups.Clear();
-
-            // If no RenderOrder was set, just render everything.
-            if (RenderOrder == null)
-            {
-                AllRenderables.Render(
-                    GlobalRenderGroupOptions, renderAction, seenRenderables, seenGroups);
-            }
-            // Otherwise, render in the order specified by the RenderOrder.
-            else
-            {
-                /* TODO: Make this faster (somehow).
-                 * 
-                 * The problem with this is that we're iterating through EVERY RenderableObject and checking
-                 * if the camera can see it, then rendering it. Ideally what we should be doing instead is
-                 * iterating through every RenderableObject returned by the QueryAABB and somehow rendering
-                 * only THOSE in order.
-                 * 
-                 * For large worlds, the number of renderables visible to the camera can be MUCH less than the
-                 * total number of RenderableObjects, which is a problem.
-                 * 
-                 * Most likely what will be done is upon setting the layer's RenderOrder, some sort of mapping
-                 * is created which assigns each RenderableObject a number indicating it's position in the sorted
-                 * list of objects to render.
-                 */
-                foreach (var item in RenderOrder.Actions)
-                {
-                    switch (item.ActionType)
-                    {
-                        case RenderOrder.RenderOrderActionType.RenderItem:
-                            HandleRenderOrderAction_RenderItem((RenderOrder.RenderItem)item, renderAction);
-                            break;
-                        case RenderOrder.RenderOrderActionType.RenderGroup:
-                            HandleRenderOrderAction_RenderGroup((RenderOrder.RenderGroup)item, renderAction);
-                            break;
-                        case RenderOrder.RenderOrderActionType.SetRenderProperties:
-                            HandleRenderOrderAction_SetRenderProperties((RenderOrder.SetRenderProperties)item);
-                            break;
-                        default:
-                            HandleUnknownRenderOrderAction(item, renderAction, isVisibleToCamera);
-                            break;
-                    }
-                }
-            }
-
+        /// <summary>
+        /// Render this layer's target. This is called immediately after all the Renderables
+        /// have been rendered, and immediately before PostRender.
+        /// </summary>
+        protected override void RenderLayerTarget()
+        {
             rp.UnlockMatrix();
             rp.UnsetRenderTarget();
             rp.ClearRenderProperties();
@@ -712,9 +395,10 @@ namespace Artemis.Engine.Graphics
             rp.Render(LayerTarget, Vector2.Zero);
 
             rp.ClearRenderProperties();
+        }
 
-            midRender = false;
-
+        protected override void PostRender()
+        {
             if (_requiresTargetTransformRecalc)
             {
                 if (_newLayerScaleType.HasValue)
@@ -725,78 +409,8 @@ namespace Artemis.Engine.Graphics
                 _newLayerScaleType = null;
                 _newUniformScaleType = null;
 
-                RecalculateTargetTransform();
+                RecalculateTargetTransform(); // sets _requiresTargetTransformRecalc to false.
             }
-
-            // Add all the items that were added mid-render.
-            //
-            // Honestly, this is kind of an unnecessary buffer for user clumsiness. Their
-            // rendering code should NOT be adding anything to a layer, that SHOULD be happening
-            // in their update code. This is just in case there's some edge case where you're
-            // forced to add an item mid-render, or in case the user isn't astute enough to
-            // realize the error in their ways.
-
-            foreach (var item in renderablesToAdd)
-            {
-                if (item.Anonymous)
-                {
-                    if (item.Name != null)
-                        AddAnonymousItem(item.Name, item.Object);
-                    else
-                        AddAnonymousItem(item.Object);
-                }
-                else
-                {
-                    AddItem(item.Name, item.Object);
-                }
-            }
-
-            renderablesToAdd.Clear();
-        }
-
-        /// <summary>
-        /// Handle a RenderItem action in the RenderOrder.
-        /// </summary>
-        /// <param name="action"></param>
-        private void HandleRenderOrderAction_RenderItem(
-            RenderOrder.RenderItem action, RenderAction renderAction)
-        {
-            var renderable = AllRenderables.GetItem(action.Name);
-            if (!(action.SkipDuplicates && seenRenderables.Contains(renderable)))
-                renderAction(renderable);
-        }
-        
-        /// <summary>
-        /// Handle a RenderGroup action in the RenderOrder.
-        /// </summary>
-        /// <param name="action"></param>
-        private void HandleRenderOrderAction_RenderGroup(
-            RenderOrder.RenderGroup action, RenderAction renderAction)
-        {
-            var group = AllRenderables.GetSubnode(action.Name);
-            group.Render(action.Options, renderAction, seenRenderables, seenGroups, action.SkipDuplicates);
-        }
-
-        /// <summary>
-        /// Handle a SetRenderProperties action in the RenderOrder.
-        /// </summary>
-        /// <param name="action"></param>
-        private void HandleRenderOrderAction_SetRenderProperties(RenderOrder.SetRenderProperties action)
-        {
-            ArtemisEngine.RenderPipeline.SetRenderProperties(
-                action.Packet, action.IgnoreDefaults, action.ApplyMatrix);
-        }
-
-        /// <summary>
-        /// Handle a RenderOrderAction that the RenderLayer couldn't identify what to do with.
-        /// Implement this if you have RenderOrderActions that aren't one of the builtin types.
-        /// </summary>
-        /// <param name="action"></param>
-        protected virtual void HandleUnknownRenderOrderAction(
-            RenderOrder.AbstractRenderOrderAction action, RenderAction renderAction, Predicate<RenderableObject> isVisibleToCamera)
-        {
-            throw new RenderOrderException(
-                String.Format("Unknown render order action '{0}'.", action));
         }
     }
 }
